% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_power.R
\name{get_power}
\alias{get_power}
\title{Get Power for a Specific N, Mean Pattern, and Repititions}
\usage{
get_power(n, a1b1, a1b2, a2b1, a2b2, alpha, reps)
}
\arguments{
\item{n}{Sample size}

\item{a1b1}{Mean in cell A1, B1}

\item{a1b2}{Mean in cell A1, B2}

\item{a2b1}{Mean in cell A2, B1}

\item{a2b2}{Mean in cell A2, B2}

\item{alpha}{Alpha level}

\item{reps}{Number of simulations to run}
}
\value{
Numeric indicating the power
}
\description{
Get Power for a Specific N, Mean Pattern, and Repititions
}
