% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means.R
\name{plot_means}
\alias{plot_means}
\title{Plot the Two-Way Interaction Mean Patterns to be Simulated}
\usage{
plot_means(a1b1, a1b2, a2b1, a2b2, a = NULL, a1 = NULL, a2 = NULL,
  b = NULL, b1 = NULL, b2 = NULL)
}
\arguments{
\item{a1b1}{Mean in cell A1, B1}

\item{a1b2}{Mean in cell A1, B2}

\item{a2b1}{Mean in cell A2, B1}

\item{a2b2}{Mean in cell A2, B2}

\item{a}{String. Name of the factor (variable) representing A (x-axis)}

\item{a1}{String. Name of the first level of Factor A}

\item{a2}{String. Name of the second level of Factor A}

\item{b}{String. Name of the factor (variable) representing B 
(different series/colors)}

\item{b1}{String. Name of the first level of Factor B}

\item{b2}{String. Name of the second level of Factor B}
}
\value{
A ggplot2 object depicting the mean patterns
}
\description{
Plot to see what pattern of means look like before one puts them
  into the simulation. This allows a user to make sure what they input to
  the simulation function matches the pattern of results they expect to see.
}
\examples{
# imagine d = .5 in condition A and d = .0 in condition B
plot_means(0, .5, 0, 0)

}
