% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_model.R
\name{sim_model}
\alias{sim_model}
\title{Simulate a Model and Extract if p < alpha}
\usage{
sim_model(n, a1b1, a1b2, a2b1, a2b2, alpha)
}
\arguments{
\item{n}{Sample size}

\item{a1b1}{Mean in cell A1, B1}

\item{a1b2}{Mean in cell A1, B2}

\item{a2b1}{Mean in cell A2, B1}

\item{a2b2}{Mean in cell A2, B2}

\item{alpha}{Alpha level}
}
\value{
Logical indicating TRUE for less than alpha, FALSE otherwise
}
\description{
Simulate a Model and Extract if p < alpha
}
